Context Menu Ray Cast Ignores Scoring/Protected Zones
